refactor: ToolRequestHandler を複数モジュールに分離
巨大な ToolRequestHandler.js (469行) を機能別に分割し、保守性と再利用性を向上。

- src/server/handlers/ToolRouter.js: ツール振り分けロジック専用クラス
- src/server/handlers/ErrorHandlers.js: エラーハンドリング専用モジュール
- src/utils/DataTransformers.js: データ変換ユーティリティ
- ToolRequestHandler.js: 軽量なオーケストレーター (40行程度)
